feat: restructure routes and corresponding template and static files
- Rename main_routes.py to stripe.py

- Move endpoints returning templates to frontend routes

- Refactor post products api route to handle to handle post requests only

Modify call to config endpoint in main.js to align with stripe routes prefix

- Modify create checkout route to align with stripe routes prefix

- Fix stripe package import in base.html
