+++ app/routes/__init__.py
@@ -1,6 +1,8 @@
-from .api_routes import api_bp
-from .main_routes import routes_bp
+from .api import api_bp
+from .stripe import stripe_bp
+from .frontend import frontend_bp
 
 def register_blueprints(app):
     app.register_blueprint(api_bp, url_prefix='/api')
-    app.register_blueprint(routes_bp, url_prefix='/')
+    app.register_blueprint(stripe_bp, url_prefix='/stripe')
+    app.register_blueprint(frontend_bp,url_prefix='/')
