+++ app/templates/base.html
@@ -6,6 +6,8 @@
         <title>Supa Tees</title>
         <!-- Import Bulma CSS -->
         <link rel="stylesheet" href="https://cdn.jsdelivr.net/npm/bulma@1.0.0/css/bulma.min.css">
+        <!-- Import Stripe -->
+        <script src="https://js.stripe.com/v3/"></script>
         <script src="{{ url_for('static', filename='main.js') }}"></script>
         <script defer src="https://use.fontawesome.com/releases/v5.14.0/js/all.js"></script>
     </head>
